修复：修复因 mall_sale_agent 表字段带来的参数问题

+++ globalshop-order-app/src/main/java/com/wangqin/globalshop/order/app/agent/controller/MallSaleAgentController.java
@@ -121,7 +121,7 @@ public class MallSaleAgentController {
      * @return
      */
     @PostMapping("/updateCommissionValue")
-    public Object updateCommissionValue(String userNo, Long commissionMode, Double commissionValue) {
+    public Object updateCommissionValue(String userNo, Integer commissionMode, Double commissionValue) {
         JsonResult result = new JsonResult();
 
         try {
